+++ rest-client-tests/test.Users.http
@@ -2,27 +2,26 @@
 GET http://localhost:4000/users HTTP/1.1
 
 ### get single user
-GET http://localhost:4000/users/c49391f4-436c-4ed9-a26c-e32fa7e72119 HTTP/1.1
+GET http://localhost:4000/users/73f0c29d-58c7-4253-a037-b5b0a315b9fc HTTP/1.1
 
-### add user
+### post user
 POST http://localhost:4000/users HTTP/1.1
 Content-Type: application/json
 
 {
-  "name": "Andrey 9",
-  "login": "Macbaren 9",
+  "name": "Andrey 10",
+  "login": "Macbaren 10",
   "password": "999"
 }
 
-### delete user
-DELETE http://localhost:4000/users/44dd8eca-f741-4977-ab88-adee842fa45e HTTP/1.1
-
 ### update user
-PUT http://localhost:4000/users/570d94f6-278b-41c2-9e48-e8cb94165e95 HTTP/1.1
+PUT http://localhost:4000/users/b0f50086-07cd-4d07-8465-643e35ba4a2a HTTP/1.1
 Content-Type: application/json
 
 {
-  "id": "123",
-  "login": "Macbaren 2",
   "password": "123457890"
 }
+### delete user
+DELETE http://localhost:4000/users/f46252be-4d64-4418-b134-96b18fd30edb HTTP/1.1
+
+
