+++ src/app.js
@@ -1,24 +1,14 @@
-// const express = require('express');
-// const swaggerUI = require('swagger-ui-express');
-// const path = require('path');
-// const YAML = require('yamljs');
-// const userRouter = require('./resources/users/user.router');
+const fastify = require('fastify')({ logger: true });
 
-// const app = express();
-// const swaggerDocument = YAML.load(path.join(__dirname, '../doc/api.yaml'));
+fastify.register(require('fastify-swagger'), {
+  exposeRoute: true,
+  routePrefix: '/docs',
+  swagger: {
+    info: { title: 'fastify-api' },
+  },
+});
+fastify.register(require('./resources/users/users.router'));
+fastify.register(require('./resources/boards/router'));
+fastify.register(require('./resources/tasks/router'));
 
-// app.use(express.json());
-
-// app.use('/doc', swaggerUI.serve, swaggerUI.setup(swaggerDocument));
-
-// app.use('/', (req, res, next) => {
-//   if (req.originalUrl === '/') {
-//     res.send('Service is running!');
-//     return;
-//   }
-//   next();
-// });
-
-// app.use('/users', userRouter);
-
-// module.exports = app;
+module.exports = fastify;
