+++ src/server.js
@@ -1,17 +1,5 @@
-const fastify = require('fastify')({ logger: true });
-
-fastify.register(require('fastify-swagger'), {
-  exposeRoute: true,
-  routePrefix: '/docs',
-  swagger: {
-    info: { title: 'fastify-api' },
-  },
-});
-fastify.register(require('./resources/users/router'));
-fastify.register(require('./resources/boards/router'));
-fastify.register(require('./resources/tasks/router'));
-
 const { PORT } = require('./common/config');
+const fastify = require('./app');
 
 const start = async () => {
   try {
